IKASAN-1337 - move package names of samples from org.ikasan to com.ikasan remove IKASAN-1337 @RunWith(SpringRunner.class) from integration tests
