+++ ikasaneip/sample/spring-boot/builder-pattern/src/main/java/com/ikasan/sample/spring/boot/builderpattern/ModuleConfig.java
@@ -1,0 +1,16 @@
+package com.ikasan.sample.spring.boot.builderpattern;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.ImportResource;
+
+
+@Configuration
+@ImportResource( {
+        "classpath:ikasan-transaction-pointcut-jms.xml",
+        "classpath:ikasan-transaction-pointcut-resubmission.xml",
+        "classpath:ikasan-transaction-pointcut-quartz.xml"
+} )
+public class ModuleConfig
+{
+
+}
