+++ ikasaneip/sample/spring-boot/builder-pattern/src/test/java/com/ikasan/sample/spring/boot/builderpattern/ModuleTestConfig.java
@@ -1,0 +1,28 @@
+package com.ikasan.sample.spring.boot.builderpattern;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.jms.annotation.EnableJms;
+import org.springframework.jms.config.JmsListenerContainerFactory;
+import org.springframework.jms.config.SimpleJmsListenerContainerFactory;
+import org.springframework.jms.core.JmsTemplate;
+
+import javax.jms.ConnectionFactory;
+
+@Configuration
+@EnableJms
+public class ModuleTestConfig
+{
+    @Bean JmsListenerContainerFactory<?> myJmsContainerFactory(ConnectionFactory connectionFactory)
+    {
+        SimpleJmsListenerContainerFactory factory = new SimpleJmsListenerContainerFactory();
+        factory.setConnectionFactory(connectionFactory);
+        return factory;
+    }
+
+    @Bean JmsTemplate jmsTemplate(ConnectionFactory connectionFactory)
+    {
+        JmsTemplate jmsTemplate = new JmsTemplate(connectionFactory);
+        return jmsTemplate;
+    }
+}
