+++ ikasaneip/sample/spring-boot/file/src/test/java/com/ikasan/sample/spring/boot/builderpattern/ApplicationTest.java
@@ -62,7 +62,6 @@ import static org.junit.Assert.assertTrue;
  * 
  * @author Ikasan Development Team
  */
-@RunWith(SpringRunner.class)
 public class ApplicationTest
 {
     public EmbeddedActiveMQBroker broker;
