+++ ikasaneip/sample/spring-boot/ftp/README.md
@@ -15,7 +15,7 @@ sample-spring-boot-ftp provides example of integration module using FTP. The mod
   * Ftp Producer - Delivers the file to FTP server
 
 ## How to construct ftpToLogFlow using builder pattern
-Check out the source code at [Application](src/main/java/org/ikasan/sample/spring/boot/builderpattern/Application.java)
+Check out the source code at [Application](src/main/java/com/ikasan/sample/spring/boot/builderpattern/Application.java)
 ```java
 public Flow getFtpToLogFlow(ModuleBuilder moduleBuilder, ComponentBuilder componentBuilder) {
   FlowBuilder ftpToLogFlowBuilder = moduleBuilder.getFlowBuilder("ftpToLogFlow");
