+++ ikasaneip/sample/spring-boot/ftp/src/main/java/com/ikasan/sample/converter/FilePayloadGeneratorConverter.java
@@ -1,0 +1,80 @@
+/*
+ * $Id$
+ * $URL$
+ *
+ * ====================================================================
+ * Ikasan Enterprise Integration Platform
+ *
+ * Distributed under the Modified BSD License.
+ * Copyright notice: The copyright for this software and a full listing
+ * of individual contributors are as shown in the packaged copyright.txt
+ * file.
+ *
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ *  - Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ *
+ *  - Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ *  - Neither the name of the ORGANIZATION nor the names of its contributors may
+ *    be used to endorse or promote products derived from this software without
+ *    specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
+ * USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ * ====================================================================
+ */
+package com.ikasan.sample.converter;
+
+import org.ikasan.filetransfer.Payload;
+import org.ikasan.filetransfer.component.DefaultPayload;
+import org.ikasan.spec.component.transformation.Converter;
+import org.quartz.JobExecutionContext;
+
+import java.util.Date;
+
+/**
+ * Get the payload content from the incoming message.
+ * Created by Ikasan Development Team
+ */
+public class FilePayloadGeneratorConverter implements Converter<JobExecutionContext,Payload>
+{
+
+    private String fileName;
+
+    @Override
+    public Payload convert(JobExecutionContext data)
+    {
+        String localFileName;
+        if(fileName==null || fileName.isEmpty()){
+            localFileName = "test"+new Date().getTime()+".txt";
+        } else
+        {
+            localFileName=fileName;
+        }
+        String message = data.getJobDetail().getKey().getGroup()+data.getJobDetail().getKey().getName()+data.getFireTime();
+        Payload payload =  new DefaultPayload(localFileName
+                ,message.getBytes());
+        payload.setAttribute("fileName",localFileName);
+        return payload;
+    }
+
+    public void setFileName(String fileName)
+    {
+        this.fileName = fileName;
+    }
+}
