--- ikasaneip/sample/spring-boot/jms/src/main/java/org/ikasan/sample/spring/boot/builderpattern/Application.java
@@ -1,15 +1,0 @@
-package org.ikasan.sample.spring.boot.builderpattern;
-
-import org.ikasan.builder.IkasanApplication;
-import org.ikasan.builder.IkasanApplicationFactory;
-
-public class Application
-{
-
-    public static void main(String[] args) throws Exception
-    {
-        IkasanApplication ikasanApplication = IkasanApplicationFactory.getIkasanApplication(args);
-
-        System.out.println("Context ready");
-    }
-}
