--- ikasaneip/sample/spring-boot/jms/src/main/java/org/ikasan/sample/spring/boot/builderpattern/SampleGeneratedException.java
@@ -1,17 +1,0 @@
-package org.ikasan.sample.spring.boot.builderpattern;
-
-/**
- * Created by majean on 09/10/2017.
- */
-public class SampleGeneratedException extends RuntimeException
-{
-    public SampleGeneratedException(Throwable e){
-        super(e);
-    }
-
-    public SampleGeneratedException(String m){
-        super(m);
-    }
-
-
-}
