+++ ikasaneip/sample/spring-boot/jms/src/test/java/com/ikasan/sample/spring/boot/builderpattern/MessageVerifierConsumer.java
@@ -1,0 +1,27 @@
+package com.ikasan.sample.spring.boot.builderpattern;
+
+import org.springframework.jms.annotation.JmsListener;
+import org.springframework.stereotype.Component;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Component
+public class MessageVerifierConsumer
+{
+    List<String> captureResults= new ArrayList<>();
+
+    @JmsListener(destination = "target", containerFactory = "myJmsContainerFactory" )
+    public void receiveMessage(String msg)
+    {
+        System.out.println("Received :" + msg);
+        captureResults.add(msg);
+
+    }
+
+    public List<String> getCaptureResults()
+    {
+        return captureResults;
+    }
+
+}
