+++ ikasaneip/sample/spring-boot/sftp/src/main/resources/sftp-to-log-component-conf.xml
@@ -7,9 +7,9 @@
     <!--  component implementations                                          -->
     <!--  ================================================================== -->
 
-    <bean id="payloadToStringConverter" class="org.ikasan.sample.converter.PayloadToStringConverter"/>
+    <bean id="payloadToStringConverter" class="com.ikasan.sample.converter.PayloadToStringConverter"/>
 
-    <bean id="filePayloadGeneratorConverter" class="org.ikasan.sample.converter.FilePayloadGeneratorConverter"/>
+    <bean id="filePayloadGeneratorConverter" class="com.ikasan.sample.converter.FilePayloadGeneratorConverter"/>
 
     <bean id="fileGeneratorScheduledConsumer" class="org.ikasan.component.endpoint.quartz.consumer.ScheduledConsumer">
         <constructor-arg ref="scheduler"/>
